archive/FrontEnd/requirements.txt
quart==0.20.0
python-socketio==5.13.0
polygon-api-client==1.12.4
black==23.1.0
uvicorn